+++ src/controllers/admin.ts
@@ -62,7 +62,7 @@ const serverStatus = async (req: Request, res: Response): Promise<Response> => {
         status: "success",
         message: "Nostrcheck-server is running.",
 		version: process.env.npm_package_version || "0.0.0",
-		uptime: format(process.uptime()),
+		uptime: format(process.uptime()), 
         ramUsage: Math.floor(process.memoryUsage().rss / 1024 / 1024),
         cpuUsage: await getCPUUsage(),
         moderationQueue: getModerationQueueLength(),