+++ src/controllers/payments.ts
@@ -6,11 +6,10 @@ import { logger } from "../lib/logger.js";
 import { ResultMessagev2 } from "../interfaces/server.js";
 import { parseAuthHeader} from "../lib/authorization.js";
 import { isModuleEnabled} from "../lib/config.js";
-import { payInvoiceFromExpenses, addBalance, getBalance, formatAccountNumber, getInvoice, calculateSatoshi, collectInvoice } from "../lib/payments/core.js";
+import { payInvoiceFromExpenses, addBalance, getBalance, formatAccountNumber, getInvoice, collectInvoice } from "../lib/payments/core.js";
 import { dbMultiSelect } from "../lib/database.js";
 import { isInvoicePaid } from "../lib/payments/core.js";
-import { amountReturnMessage, invoiceReturnMessage } from "../interfaces/payments.js";
-import { getDomainInfo } from "../lib/domains.js";
+import { invoiceReturnMessage } from "../interfaces/payments.js";
 import { setAuthCookie } from "../lib/frontend.js";
 import { isIpAllowed } from "../lib/security/ips.js";
 
@@ -175,50 +174,6 @@ const getInvoiceStatus = async (req: Request, res: Response): Promise<Response> 
     return res.status(200).send(result);
 }
 
-const calculateObjectAmount = async (req: Request, res: Response): Promise<Response> => {
-
-    // Check if the request IP is allowed
-	const reqInfo = await isIpAllowed(req);
-	if (reqInfo.banned == true) {
-		logger.info(`calculateObjectAmount - Attempt to access ${req.path} with unauthorized IP:`, reqInfo.ip);
-		return res.status(403).send({"status": "error", "message": reqInfo.comments});
-	}
-
-    // Check if current module is enabled
-    if (!isModuleEnabled("admin", app)) {
-        logger.info(`calculateObjectAmount - Attempt to access a non-active module: admin | IP:`, reqInfo.ip);
-        return res.status(403).send({"status": "error", "message": "Module is not enabled"});
-    }
-
-    logger.info(`calculateObjectAmount - Request from:`, req.hostname, "|", reqInfo.ip);
-    res.setHeader('Content-Type', 'application/json');
-
-    // Check if the request has the required parameters
-    if (req.body.size === undefined || req.body.size === null) {
-        const result : ResultMessagev2 = {
-            status: "error",
-            message: "Invalid parameters"
-            };
-        logger.error(`calculateObjectAmount - Invalid parameters | ${reqInfo.ip}`);
-        return res.status(400).send(result);
-    }
-
-    const size = req.body.size;
-    const domain = req.body.domain || "";
-    const domainInfo = await getDomainInfo(domain)
-
-    const satoshi = await calculateSatoshi(domain != "" ? 'registered': 'mediafiles', size, domainInfo != "" ? domainInfo.maxsatoshi : app.get("config.payments")["satoshi"]["mediaMaxSatoshi"]);
-
-    const result : amountReturnMessage = {
-        status: "success",
-        message: "Calculated satoshi successfully",
-        amount: satoshi
-        };
-    logger.info(`calculateObjectAmount - Calculated satoshi successfully: ${satoshi}, size: ${size}, domain: ${domain} | ${reqInfo.ip}`);
-    return res.status(200).send(result);
-    
-}
-
 const getBalanceUser = async (req: Request, res: Response): Promise<Response> => {
 
     // Check if the request IP is allowed
@@ -273,5 +228,4 @@ export {
     addBalanceUser,
     getBalanceUser,
     getInvoiceStatus,
-    calculateObjectAmount
 }