+++ src/interfaces/domains.ts
@@ -1,8 +1,0 @@
-type domainInfo = {
-    requireinvite: boolean;
-    requirepayment: boolean;
-    maxsatoshi: number;
-};
-
-
-export { domainInfo };