+++ src/lib/payments/LUD06.ts
@@ -1,10 +1,10 @@
 import axios from 'axios';
 import * as bolt11 from 'bolt11';
-import { emptyInvoice, invoice } from '../../interfaces/payments.js';
+import { emptyInvoice, Invoice } from '../../interfaces/payments.js';
 import { logger } from '../logger.js';
 import { getNewDate } from '../utils.js';
 
-async function generateLUD06Invoice(lnurl: string, amount: number): Promise<invoice> {
+async function generateLUD06Invoice(lnurl: string, amount: number): Promise<Invoice> {
 
   if (lnurl === '' || amount === 0) {return emptyInvoice;}